+++ src/valuta/contrib/django_integration/templatetags/valuta_tags.py
@@ -1,9 +1,10 @@
 from decimal import Decimal
-from typing import Union
+from typing import Union, Optional
 
 from django import template
 
-from ....shortcuts import convert_to_currency_units
+from ....constants import DEFAULT_DISPLAY_FORMAT
+from ....shortcuts import convert_to_currency_units, display_in_currency_units
 
 register = template.Library()
 
@@ -13,6 +14,8 @@ __license__ = "GPL-2.0-only OR LGPL-2.1-or-later"
 __all__ = (
     "filter_convert_to_currency_units",
     "tag_convert_to_currency_units",
+    "filter_display_in_currency_units",
+    "tag_display_in_currency_units",
 )
 
 
@@ -28,3 +31,21 @@ def tag_convert_to_currency_units(
     value: int, currency_code: str
 ) -> Union[int, float, Decimal, None]:
     return convert_to_currency_units(currency_code, value)
+
+
+@register.filter(name="display_in_currency_units")
+def filter_display_in_currency_units(
+    value: int,
+    currency_code: str,
+    format: Optional[str] = DEFAULT_DISPLAY_FORMAT,
+) -> Union[str, None]:
+    return display_in_currency_units(currency_code, value, format)
+
+
+@register.simple_tag(name="display_in_currency_units")
+def tag_display_in_currency_units(
+    value: int,
+    currency_code: str,
+    format: Optional[str] = DEFAULT_DISPLAY_FORMAT,
+) -> Union[str, None]:
+    return display_in_currency_units(currency_code, value, format)