+++ src/valuta/tests/test_shortcuts.py
@@ -2,7 +2,7 @@ import unittest
 
 from ..currencies import TND, EUR, MRU, VND, UGX
 from ..exceptions import InvalidCurrency
-from ..shortcuts import convert_to_currency_units
+from ..shortcuts import convert_to_currency_units, display_in_currency_units
 
 
 __author__ = "Artur Barseghyan"
@@ -13,6 +13,10 @@ __all__ = ("TestShortcuts",)
 
 class TestShortcuts(unittest.TestCase):
     """Test shortcuts."""
+
+    # ***********************************************************
+    # *************** Convert to currency units *****************
+    # ***********************************************************
 
     def test_convert_to_currency_units(self):
         """Test `convert_to_currency_units` with valid codes."""
@@ -64,3 +68,73 @@ class TestShortcuts(unittest.TestCase):
         with self.assertRaises(InvalidCurrency):
             with self.subTest("currency_code='None'"):
                 self._test_currency_code_is_invalid(fail_silently=False)
+
+    # ***********************************************************
+    # *************** Display in currency units *****************
+    # ***********************************************************
+
+    def test_display_in_currency_units(self):
+        """Test `display_in_currency_units` with valid codes."""
+        self.assertEqual(display_in_currency_units(UGX.uid, 10), "10")
+
+        self.assertEqual(display_in_currency_units(MRU.uid, 5), "1.00")
+
+        self.assertEqual(display_in_currency_units(VND.uid, 10), "1")
+
+        self.assertEqual(display_in_currency_units(EUR.uid, 100), "1.00")
+
+        self.assertEqual(display_in_currency_units(TND.uid, 1_000), "1.000")
+
+    def _display_test_currency_code_is_none(self, fail_silently: bool = True):
+        self.assertEqual(
+            display_in_currency_units(
+                None, 1_000, fail_silently=fail_silently
+            ),
+            None,
+        )
+
+    def _display_test_currency_code_is_empty_string(
+        self, fail_silently: bool = True
+    ):
+        self.assertEqual(
+            display_in_currency_units("", 1_000, fail_silently=fail_silently),
+            None,
+        )
+
+    def _display_test_currency_code_is_invalid(
+        self, fail_silently: bool = True
+    ):
+        self.assertEqual(
+            display_in_currency_units(
+                "None", 1_000, fail_silently=fail_silently
+            ),
+            None,
+        )
+
+    def test_test_display_in_currency_units_invalid_currency_code(self):
+        """Test `convert_to_currency_units` with invalid codes."""
+
+        with self.subTest("currency_code=None"):
+            self._display_test_currency_code_is_none()
+
+        with self.subTest("currency_code=''"):
+            self._display_test_currency_code_is_empty_string()
+
+        with self.subTest("currency_code='None'"):
+            self._display_test_currency_code_is_invalid()
+
+        with self.assertRaises(InvalidCurrency):
+            with self.subTest("currency_code=None"):
+                self._display_test_currency_code_is_none(fail_silently=False)
+
+        with self.assertRaises(InvalidCurrency):
+            with self.subTest("currency_code=''"):
+                self._display_test_currency_code_is_empty_string(
+                    fail_silently=False
+                )
+
+        with self.assertRaises(InvalidCurrency):
+            with self.subTest("currency_code='None'"):
+                self._display_test_currency_code_is_invalid(
+                    fail_silently=False
+                )